Move files to a Kibiter folder
