+++ src/ui/public/kibiter/menu/custom_menu.js
