+++ src/ui/public/kibiter/menu/get_metadashboard.js
